core: fix MSVC /WX warnings (C4334 32-bit shift, C4244 narrowing fill)
Two more CI-only failures, this time from the Windows CI's actual MSVC
build (windows-latest uses cl.exe via MSBuild, not the MinGW toolchain
used for local development - different warning set entirely, which is
why these weren't caught locally):

- pio_assembler.cpp: `1 << delay_bits` / `1 << ss_bits` are 32-bit int
  shifts immediately widened to compare against an int64_t (`delay`/
  `*side`) - MSVC's C4334 flags this as a likely 32-vs-64-bit mistake
  even though delay_bits/ss_bits are always small enough here for it
  not to matter in practice. Made the shift itself 64-bit
  (`std::int64_t{1} << ...`) to match the comparison's type exactly.

- st7789.cpp: `std::fill(fb_.begin(), fb_.end(), 0u)` clears a
  vector<uint16_t> with an `unsigned int` (0u) fill value - MSVC's
  standard library implementation narrows that per-element, triggering
  C4244. Passed `std::uint16_t{0}` instead, matching the container's
  own element type.

Verified against the actual MSVC 19.51 toolchain (VS 18 Community,
matching the CI runner) with /W4 /permissive- /WX, not just guessed at
from the log - configured a separate build-msvc/ (gitignored) and ran
the full suite: 46/46 passing, zero warnings-as-errors. Also confirmed
clean on the MinGW build used for local development.

+++ src/core/thumb_decode.cpp
@@ -227,7 +227,7 @@ DecodedInstr decode_misc16(std::uint16_t instr) {
         case 0b0100:
         case 0b0101: {  // PUSH
             d.op = Mnemonic::PUSH;
-            d.register_list = static_cast<std::uint16_t>((bit(instr, 8) << 14) | (instr & 0xFFu));
+            d.register_list = static_cast<std::uint16_t>((static_cast<unsigned>(bit(instr, 8)) << 14) | (instr & 0xFFu));
             return d;
         }
         case 0b0110: {  // CPS (only valid misc encoding in this row for ARMv6-M)
@@ -250,7 +250,7 @@ DecodedInstr decode_misc16(std::uint16_t instr) {
         case 0b1100:
         case 0b1101: {  // POP
             d.op = Mnemonic::POP;
-            d.register_list = static_cast<std::uint16_t>((bit(instr, 8) << 15) | (instr & 0xFFu));
+            d.register_list = static_cast<std::uint16_t>((static_cast<unsigned>(bit(instr, 8)) << 15) | (instr & 0xFFu));
             return d;
         }
         case 0b1110:  // BKPT
